+++ package.json
@@ -24,7 +24,9 @@
     "morgan": "^1.10.0",
     "pg": "^8.5.1",
     "uuid": "^8.3.2",
-    "winston": "^3.3.3"
+    "valid-url": "^1.0.9",
+    "winston": "^3.3.3",
+    "xss": "^1.0.8"
   },
   "devDependencies": {
     "chai": "^4.3.0",
